requirements.txt
accelerate==1.7.0
addict==2.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
bitsandbytes==0.46.0
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
braceexpand==0.1.7
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
comm==0.2.2
confection==0.1.5
ConfigArgParse==1.7.1
contexttimer==0.3.3
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
dash==3.0.4
decorator==5.2.1
decord==0.6.0
diffusers==0.16.0
distlib==0.3.9
easydict==1.9
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fairscale==0.4.4
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
fonttools==4.58.2
fsspec==2025.5.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
h5py==3.14.0
hf-xet==1.1.3
huggingface-hub==0.25.2
identify==2.6.12
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.10
ipython==8.18.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kaggle==1.7.4.5
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
murmurhash==1.0.13
narwhals==1.42.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
nodeenv==1.9.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.11.0.86
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==24.2
pandas==2.3.0
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pillow==11.2.1
pip==25.1
platformdirs==4.3.8
plotly==6.1.2
portalocker==3.1.1
pre_commit==4.2.0
preshed==3.0.10
proglog==0.1.12
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
retrying==1.3.4
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
salesforce-lavis==1.0.1
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
soundfile==0.13.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
streamlit==1.45.1
sympy==1.13.1
tenacity==9.1.2
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.33.2
triton==3.2.0
typer==0.16.0
typing_extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
zipp==3.23.0
